# Use `-Ofast` compiler optimization