@jtraglia 2023-02-10
Actually, nevermind. This enables optimizations which violate requirements of C standard for floating-point semantics. I don't think this would be an issue, but not worth the risk for <1ms of performance.